apps/translate_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
grpcio
grpcio-tools
pika
googletrans==4.0.0-rc1